Some refactorings to make room for new features
